CSED490F_DDP_DALI_Training-main/requirements.txt
numpy<2
nvidia-dali-cuda120
pynvml
pandas
matplotlib
nvtx